추가 : 그래프 구조 , KioskGui
1. 그래프 구조 추가
2. KioskGui추가
  -> 호버링 테스트, 클릭 인식 테스트 완료

+++ eternal_return/src/main/java/org/EternalReturn/Util/Gui/bossbarGui/Model/BComponent.java
@@ -51,11 +51,11 @@ public class BComponent{
         return component;
     }
 
-    public int getSizeX() {
+    public float getSizeX() {
         return sizeX;
     }
 
-    public int getSizeY() {
+    public float getSizeY() {
         return sizeY;
     }
